Add local pillar requirement to integration tests
NOTE: This needs to be broadcast and updated in contributing docs,
probably in the quickstart in the README as well. This allows us to
start moving away from the test_value_scraper, instead we only have to
provide a path to some file containing the pillar data, e.g.

saltext.vmware:
  vcenter:
    ...

Or whatever.